+++ doc/api/actions/move_by.md
@@ -6,6 +6,14 @@ Move the Magnebot forward or backward by a given distance.
 
 ***
 
+## Class Variables
+
+| Variable | Type | Description |
+| --- | --- | --- |
+| `JOINT_ORDER` | Dict[Arm, List[ArmJoint]] | The order in which joint angles will be set. |
+
+***
+
 ## Fields
 
 - `status` [The current status of the action.](../action_status.md) By default, this is `ongoing` (the action isn't done).
@@ -66,7 +74,7 @@ Evaluate an action per-frame to determine whether it's done.
 | static |  MagnebotStatic |  | [The static Magnebot data.](../magnebot_static.md) |
 | dynamic |  MagnebotDynamic |  | [The dynamic Magnebot data.](../magnebot_dynamic.md) |
 
-_Returns:_  Tuple: A list of commands to send to the build to continue the action.
+_Returns:_  A list of commands to send to the build to continue the action.
 
 #### get_end_commands
