+++ doc/api/actions/wheel_motion.md
@@ -6,6 +6,14 @@ Abstract base class for a motion action involving the Magnebot's wheels.
 
 ***
 
+## Class Variables
+
+| Variable | Type | Description |
+| --- | --- | --- |
+| `JOINT_ORDER` | Dict[Arm, List[ArmJoint]] | The order in which joint angles will be set. |
+
+***
+
 ## Fields
 
 - `status` [The current status of the action.](../action_status.md) By default, this is `ongoing` (the action isn't done).
@@ -31,6 +39,8 @@ Abstract base class for a motion action involving the Magnebot's wheels.
 | previous |  Action  | None | The previous action, if any. |
 
 #### get_initialization_commands
+
+_(Final)_
 
 **`self.get_initialization_commands(resp, static, dynamic, image_frequency)`**
 
@@ -59,7 +69,7 @@ Evaluate an action per-frame to determine whether it's done.
 | static |  MagnebotStatic |  | [The static Magnebot data.](../magnebot_static.md) |
 | dynamic |  MagnebotDynamic |  | [The dynamic Magnebot data.](../magnebot_dynamic.md) |
 
-_Returns:_  Tuple: An `ActionStatus` describing whether the action is ongoing, succeeded, or failed; A list of commands to send to the build if the action is ongoing.
+_Returns:_  A list of commands to send to the build if the action is ongoing.
 
 #### get_end_commands
 
@@ -99,5 +109,5 @@ Evaluate an action per-frame to determine whether it's done.
 | static |  MagnebotStatic |  | [The static Magnebot data.](../magnebot_static.md) |
 | dynamic |  MagnebotDynamic |  | [The dynamic Magnebot data.](../magnebot_dynamic.md) |
 
-_Returns:_  Tuple: A list of commands to send to the build to continue the action.
+_Returns:_  A list of commands to send to the build to continue the action.
